move out time and graph
