+++ pkg/chaosdaemon/graph.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package utils
+package chaosdaemon
 
 // Edge represents an edge in graph
 type Edge struct {
